Always resolve to the first response
I didn't write this app originally so I don't understand the intended behavior. It looks me like maybe the idea was to record all the API calls so that the first request returned the response to the first request and the second request returned the second to the second time the same call was made, but that isn't the behavior I want. I just want this to record the response to an API call the first time it is called and then use that response every other time that endpoint is called.